Keyword_Optimization/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
PyPDF2>=2.0.0
groq>=0.3.0
python-multipart>=0.0.5
pydantic>=1.8.0
